backend/requirements.txt
fastapi
uvicorn
prisma
python-dotenv
pyjwt
passlib[bcrypt]
python-multipart
google-generativeai
requests
asyncssh
reportlab
markdown
